gbm-noticeboard/requirements.txt
Django==2.0.2
django-environ==0.4.*
psycopg2
Pillow
# django crud adminlte2
django-crispy-forms
easy-thumbnails
django-image-cropping
djangoajax
django-Select2
django-cruds-adminlte
# django-import-export
django-import-export
